% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc.R, R/foreach.R, R/getsyms.R
\name{foreach-ext}
\alias{foreach-ext}
\alias{makeAccum}
\alias{accumulate}
\alias{getResult}
\alias{getErrorValue}
\alias{getErrorIndex}
\alias{accumulate.iforeach}
\alias{getResult.iforeach}
\alias{getErrorValue.iforeach}
\alias{getErrorIndex.iforeach}
\alias{accumulate.ixforeach}
\alias{getResult.ixforeach}
\alias{getErrorValue.ixforeach}
\alias{getErrorIndex.ixforeach}
\alias{accumulate.ifilteredforeach}
\alias{getResult.ifilteredforeach}
\alias{getErrorValue.ifilteredforeach}
\alias{getErrorIndex.ifilteredforeach}
\alias{getexports}
\title{foreach extension functions}
\usage{
makeAccum(it)

accumulate(obj, result, tag, ...)

getResult(obj, ...)

getErrorValue(obj, ...)

getErrorIndex(obj, ...)

\method{accumulate}{iforeach}(obj, result, tag, ...)

\method{getResult}{iforeach}(obj, ...)

\method{getErrorValue}{iforeach}(obj, ...)

\method{getErrorIndex}{iforeach}(obj, ...)

\method{accumulate}{ixforeach}(obj, result, tag, ...)

\method{getResult}{ixforeach}(obj, ...)

\method{getErrorValue}{ixforeach}(obj, ...)

\method{getErrorIndex}{ixforeach}(obj, ...)

\method{accumulate}{ifilteredforeach}(obj, result, tag, ...)

\method{getResult}{ifilteredforeach}(obj, ...)

\method{getErrorValue}{ifilteredforeach}(obj, ...)

\method{getErrorIndex}{ifilteredforeach}(obj, ...)

getexports(ex, e, env, good = character(0), bad = character(0))
}
\arguments{
\item{it}{foreach iterator.}

\item{obj}{foreach iterator object.}

\item{result}{task result to accumulate.}

\item{tag}{tag of task result to accumulate.}

\item{...}{unused.}

\item{ex}{call object to analyze.}

\item{e}{local environment of the call object.}

\item{env}{exported environment in which call object will be evaluated.}

\item{good}{names of symbols that are being exported.}

\item{bad}{names of symbols that are not being exported.}
}
\description{
These functions are used to write parallel backends for the \code{foreach}
package.  They should not be used from normal scripts or packages that use
the \code{foreach} package.
}
\section{Note}{

These functions are likely to change in future versions of the
\code{foreach} package.  When they become more stable, they will
be documented.
}

\keyword{utilities}
