% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setDoSeq.R
\name{setDoSeq}
\alias{setDoSeq}
\title{setDoSeq}
\usage{
setDoSeq(fun, data = NULL, info = function(data, item) NULL)
}
\arguments{
\item{fun}{A function that implements the functionality of \verb{\%dopar\%}.}

\item{data}{Data to be passed to the registered function.}

\item{info}{Function that retrieves information about the backend.}
}
\value{
(invisible) The previously registered \code{DoSeq} settings.
}
\description{
The \code{setDoSeq} function is used to register a sequential backend with the
foreach package.  This isn't normally executed by the user.  Instead, packages
that provide a sequential backend provide a function named \code{registerDoSeq}
that calls \code{setDoSeq} using the appropriate arguments.
}
\seealso{
\code{\link{\%dopar\%}}
}
\keyword{utilities}
