% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.R
\name{registerDoSEQ}
\alias{registerDoSEQ}
\title{registerDoSEQ}
\usage{
registerDoSEQ()
}
\description{
The \code{registerDoSEQ} function is used to explicitly register
a sequential parallel backend with the foreach package.
This will prevent a warning message from being issued if the
\code{\%dopar\%} function is called and no parallel backend has
been registered.
}
\examples{
# specify that \%dopar\% should run sequentially
registerDoSEQ()
}
\seealso{
\code{\link[doSNOW:registerDoSNOW]{doSNOW::registerDoSNOW}}
}
\keyword{utilities}
